# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MovementOperation": {
        "example": {
          "location": "ALM-CENTRAL",
          "movement_source": "SCAN",
          "product_id": 1,
          "quantity": 5
        },
        "properties": {
          "location": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "movement_source": {
            "$ref": "#/components/schemas/Source"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "product_id",
          "quantity",
          "location",
          "movement_source"
        ],
        "title": "MovementOperation",
        "type": "object"
      },
      "MovementResponse": {
        "example": {
          "created_at": "2026-02-17T10:20:00Z",
          "delta": 5,
          "id": 100,
          "location": "ALM-CENTRAL",
          "location_id": 1,
          "movement_source": "SCAN",
          "movement_type": "IN",
          "product_id": 1,
          "quantity": 5,
          "user_id": 2
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delta": {
            "title": "Delta",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "movement_source": {
            "$ref": "#/components/schemas/Source"
          },
          "movement_type": {
            "$ref": "#/components/schemas/MovementType"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "transfer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "product_id",
          "quantity",
          "delta",
          "movement_type",
          "movement_source",
          "user_id",
          "location_id",
          "location",
          "created_at"
        ],
        "title": "MovementResponse",
        "type": "object"
      },
...
      "MovementType": {
        "enum": [
          "IN",
          "OUT",
          "ADJUST"
        ],
        "title": "MovementType",
        "type": "string"
      },
      "MovementWithStockResponse": {
        "properties": {
          "movement": {
            "$ref": "#/components/schemas/MovementResponse"
          },
          "stock": {
            "$ref": "#/components/schemas/StockResponse"
          }
        },
        "required": [
          "stock",
          "movement"
        ],
        "title": "MovementWithStockResponse",
        "type": "object"
      },
...
      "Source": {
        "enum": [
          "SCAN",
          "MANUAL"
        ],
        "title": "Source",
        "type": "string"
      },
      "StockCreate": {
        "example": {
          "location": "ALM-CENTRAL",
          "product_id": 1,
          "quantity": 10
        },
        "properties": {
          "location": {
            "examples": [
              "ALM-CENTRAL"
            ],
            "maxLength": 100,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "product_id": {
            "examples": [
              1
            ],
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "examples": [
              10
            ],
            "minimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "product_id",
          "location",
          "quantity"
        ],
        "title": "StockCreate",
        "type": "object"
      },
      "StockListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StockResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "StockListResponse",
        "type": "object"
      },
      "StockResponse": {
        "example": {
          "created_at": "2026-02-17T10:00:00Z",
          "id": 7,
          "location": "ALM-CENTRAL",
          "product_id": 1,
          "quantity": 10,
          "updated_at": "2026-02-17T10:10:00Z"
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "location": {
            "examples": [
              "ALM-CENTRAL"
            ],
            "maxLength": 100,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "product_id": {
            "examples": [
              1
            ],
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "examples": [
              10
            ],
            "minimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "location",
          "quantity",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "StockResponse",
        "type": "object"
      },
      "StockUpdate": {
        "example": {
          "location": "ALM-NORTE",
          "quantity": 25
        },
        "properties": {
          "location": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "quantity": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          }
        },
        "title": "StockUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Sistema Inventariado Sensores",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/movements/out": {
      "post": {
        "operationId": "movement_out_movements_out_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovementOperation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementWithStockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Stock insuficiente"
                }
              }
            },
            "description": "Regla de negocio invalida"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Movement Out",
        "tags": [
          "movements"
        ]
      }
    },
...
    "/stocks/": {
      "get": {
        "operationId": "list_stocks_stocks__get",
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "order_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "id",
              "title": "Order By"
            }
          },
          {
            "in": "query",
            "name": "order_dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "asc",
              "title": "Order Dir"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "order_by debe ser 'id'"
                }
              }
            },
            "description": "Filtro u ordenacion invalida"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Stocks",
        "tags": [
          "stocks"
        ]
      },
      "post": {
        "operationId": "create_stock_stocks__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Producto no existe"
                }
              }
            },
            "description": "Producto inexistente o stock duplicado en ubicacion"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Stock",
        "tags": [
          "stocks"
        ]
      }
    },
    "/stocks/{stock_id}": {
      "get": {
        "operationId": "get_stock_stocks__stock_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_id",
            "required": true,
            "schema": {
              "title": "Stock Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Stock no encontrado"
                }
              }
            },
            "description": "Stock no encontrado"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Stock",
        "tags": [
          "stocks"
        ]
      },
      "patch": {
        "operationId": "update_stock_stocks__stock_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "stock_id",
            "required": true,
            "schema": {
              "title": "Stock Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Ya existe stock para esta ubicacion"
                }
              }
            },
            "description": "Ubicacion duplicada para el mismo producto"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Stock no encontrado"
                }
              }
            },
            "description": "Stock no encontrado"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Stock",
        "tags": [
          "stocks"
        ]
      }
    },
...
  }
}